The baseline JIT crashes when compiling "(1,1)/1" https://bugs.webkit.org/show_bug.cgi?id=157933
Reviewed by Benjamin Poulain.

op_div in the baseline JIT needed to better handle when both the lhs
and rhs are constants. It needs to make sure to load either the lhs or
the rhs into a register since the div generator can't handle both
the lhs and rhs being constants.

* jit/JITArithmetic.cpp:
(JSC::JIT::emit_op_div):
* tests/stress/jit-gracefully-handle-double-constants-in-math-operators.js: Added.
(assert):
(test):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201301 268f45cc-cd09-0410-ab3c-d52691b4dbfc