Only set overflow:hidden on -webkit-media-controls when the placeholder is showing. https://bugs.webkit.org/show_bug.cgi?id=157975
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/mac/controls-panel-not-clipped-out.html

overflow:hidden was added to clip out content contained in -webkit-media-controls
that sticks out of the bounds of that container when the inlinePlaybackPlaceholder
is showing. However, that style also clips out things that we want to show when the
inlinePlaybackPlaceholder is not showing. For example, the volume slider or the
controls panel when the video element's height is too short.

To fix that, we only set overflow:hidden on -webkit-media-controls when the
inlinePlaybackPlaceholder is visible.

* Modules/mediacontrols/mediaControlsApple.css:
(::-webkit-media-controls):
(::-webkit-media-controls.placeholder-showing):
Set overflow:hidden only when the placeholder is showing.

* Modules/mediacontrols/mediaControlsApple.js:
(Controller.prototype.handlePresentationModeChange):
Add the placeholderShowing class to -webkit-media-controls only when the
inlinePlaybackPlaceholder is not hidden.

LayoutTests:

This is a Mac only test. The media controls panel should still be visible
even if the media element's height is very short.

* TestExpectations:
* media/mac/controls-panel-not-clipped-out-expected.html: Added.
* media/mac/controls-panel-not-clipped-out.html: Added.
* platform/mac/TestExpectations:

* platform/mac/media/media-document-audio-repaint-expected.txt:
* platform/mac/media/video-zoom-controls-expected.txt:
Rebaseline tests.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201355 268f45cc-cd09-0410-ab3c-d52691b4dbfc