[css-grid] CRASH when getting the computed style of a grid with only absolutely positioned children https://bugs.webkit.org/show_bug.cgi?id=158537
Reviewed by Darin Adler.

Source/WebCore:

Absolute positioning occurs after layout of the grid and its in-flow contents, and does not
contribute to the sizing of any grid tracks or affect the size/configuration of the grid in
any way. This means that we should treat as empty any grid whose only children are
absolutely positioned items.

Since r201510 empty grids are no longer internally represented by a 1x1 matrix. As we were
not considering grids-with-only-absolutely-positioned-children as empty, we were trying to
access some invalid position in the internal representation of the grid triggering an ASSERT
in debug builds and a crash in release.

Test: fast/css-grid-layout/grid-only-abspos-item-computed-style-crash.html

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::valueForGridTrackList):

LayoutTests:

* fast/css-grid-layout/grid-only-abspos-item-computed-style-crash-expected.txt: Added.
* fast/css-grid-layout/grid-only-abspos-item-computed-style-crash.html: Added.
* fast/css-grid-layout/grid-template-shorthand-get-set-expected.txt: Adjusted, grid-template
does create explicit tracks so we should return 0px instead of none.
* fast/css-grid-layout/grid-template-shorthand-get-set.html: Ditto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201919 268f45cc-cd09-0410-ab3c-d52691b4dbfc