Make sure HTTPHeaderMap gets a move constructor / assignment operator https://bugs.webkit.org/show_bug.cgi?id=158695 <rdar://problem/26729511>
Reviewed by Alex Christensen.

Make sure HTTPHeaderMap gets a move constructor / assignment operator.
It was not getting an implicit one because of its user-declared
destructor. This patch drops the user-declared destructor so that
HTTPHeaderMap now gets an implicit move constructor / assignment
operator.

Not having a move constructor / assignment operator is an issue because
we rely on HTTPHeaderMap::isolatedCopy() / WTFMove() since r201623 to
pass HTTPHeaderMap across thread.

* platform/network/HTTPHeaderMap.cpp:
(WebCore::HTTPHeaderMap::~HTTPHeaderMap): Deleted.
* platform/network/HTTPHeaderMap.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@201994 268f45cc-cd09-0410-ab3c-d52691b4dbfc