Remove hasStaticPropertyTable (part 2: JSPluginElement) https://bugs.webkit.org/show_bug.cgi?id=158431
Reviewed by Chris Dumez.

All uses of hasStaticPropertyTable flag generated by bindings are wrong.

The check in pluginElementCustomGetOwnPropertySlot was somewhat dubious in the
first place (for types with static properties it would give precedence to both
static and also property storage properties; for types without static properties
it would check neither - an odd asymetry in the case of values in the storage
array, and was depending on an implementation detail that could change).

This is all now redundant anyway. None of these types have static properties.
All properties are now corretcly on the prototype (which is handled appropriately
below). This is just dead code.

* bindings/js/JSPluginElementFunctions.h:
(WebCore::pluginElementCustomGetOwnPropertySlot):
    - remove dead code.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202029 268f45cc-cd09-0410-ab3c-d52691b4dbfc