Revert part of r196034 https://bugs.webkit.org/show_bug.cgi?id=158805 rdar://problem/26788138
Reviewed by Dan Bernstein.

* NetworkProcess/NetworkLoad.cpp:
(WebKit::NetworkLoad::~NetworkLoad):
If the WebResourceLoader was destroyed and received a canAuthenticateAgainstProtectionSpace
but did not send a continueCanAuthenticateAgainstProtectionSpace answer because there's no
core loader, then the NetworkLoad will be destroyed.  When this happens, we still need to call
the callback for the challenge.
* WebProcess/Network/WebResourceLoader.cpp:
(WebKit::WebResourceLoader::canAuthenticateAgainstProtectionSpace):
If there's no core loader, we can't send IPC.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202119 268f45cc-cd09-0410-ab3c-d52691b4dbfc