WebRTC: Replace RTCPeerConnection custom constructor with a JS built-in constructor https://bugs.webkit.org/show_bug.cgi?id=158832
Reviewed by Eric Carlson and Youenn Fablet.

Use a JS built-in constructor instead of a custom constructor. This makes it easier to
initialize private fields for functions implemented as JS built-ins. The constructor
behavior is in need of updating, but that is left to a follow-up change [1].

[1] http://webkit.org/b/158936
No change in behavior.

* CMakeLists.txt:
* Modules/mediastream/RTCPeerConnection.cpp:
(WebCore::RTCPeerConnection::create):
(WebCore::RTCPeerConnection::RTCPeerConnection):
(WebCore::RTCPeerConnection::~RTCPeerConnection):
(WebCore::RTCPeerConnection::initializeWith):
* Modules/mediastream/RTCPeerConnection.h:
* Modules/mediastream/RTCPeerConnection.idl:
* Modules/mediastream/RTCPeerConnection.js:
(initializeRTCPeerConnection):
Add JS built-in constructor function.
* WebCore.xcodeproj/project.pbxproj:
* bindings/js/JSRTCPeerConnectionCustom.cpp: Removed.
(WebCore::constructJSRTCPeerConnection): Deleted.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202337 268f45cc-cd09-0410-ab3c-d52691b4dbfc