WebRTC: Add support for the negotiationneeded event in MediaEndpointPeerConnection https://bugs.webkit.org/show_bug.cgi?id=158985
Source/WebCore:

Reviewed by Eric Carlson.

Implement MediaEndpointPeerConnection's isNegotiationNeeded, markAsNeedingNegotiation and
clearNegotiationNeededState functions. The calls to these functions are already up-to-date.

Test: fast/mediastream/RTCPeerConnection-more-media-to-negotiate.html

* Modules/mediastream/MediaEndpointPeerConnection.cpp:
(WebCore::MediaEndpointPeerConnection::markAsNeedingNegotiation):
* Modules/mediastream/MediaEndpointPeerConnection.h:
* Modules/mediastream/RTCPeerConnection.cpp:
(WebCore::RTCPeerConnection::scheduleNegotiationNeededEvent):

LayoutTests:

Add tests for the RTCPeerConnection negotiationneeded event.

Reviewed by Eric Carlson.

* fast/mediastream/RTCPeerConnection-more-media-to-negotiate-expected.txt: Added.
* fast/mediastream/RTCPeerConnection-more-media-to-negotiate.html: Added.
Verify that a negotiationneeded event is fired when not all local media can be included in
an answer. (The answerer cannot reply with more m-lines than the offer had to begin with.)
* fast/mediastream/RTCPeerConnection-onnegotiationneeded-expected.txt:
* fast/mediastream/RTCPeerConnection-onnegotiationneeded.html:
Verify that a negotiationneeded event is fired when tracks are added and removed from an
RTCPeerConnection.
* platform/gtk/TestExpectations:
Unskip fast/mediastream/RTCPeerConnection-onnegotiationneeded.html.
* platform/mac/TestExpectations:
The mac port does not build with WEB_RTC enabled yet.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202339 268f45cc-cd09-0410-ab3c-d52691b4dbfc