Long spin editing text at top of message containing Reader version of web page with many GIFs https://bugs.webkit.org/show_bug.cgi?id=159444 <rdar://problem/26790386>
Reviewed by Sam Weinig.

* editing/cocoa/HTMLConverter.mm:
(fileWrapperForElement):
Instead of looking up the image's data in the cache by URL, just use the
CachedImage on the HTMLImageElement. There are situations (which seem to involve
cloning the DOM then having the cloned DOM get garbage collected) where the image
can be removed from the cache, but still be live in the document.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202863 268f45cc-cd09-0410-ab3c-d52691b4dbfc