Crash due to HTMLMediaElement at JavaScriptCore: JSC::JSLockHolder::JSLockHolder https://bugs.webkit.org/show_bug.cgi?id=159517 <rdar://problem/27221109>
Reviewed by Eric Carlson.

When WebKit on iOS gets a notification that the UIProcess has been backgrounded, it sends an
interruption event to the WebProcess to pause any playing HTMLMediaElements. When the
elements which get this interruption have pending promises created during a previous call to
play(), these promises get rejected.

However, if the HTMLMediaElement's document has already been destroyed, the pending Promises
are in an inconsistent state: their script execution context (the document) has been
destroyed, leading to the crash in JSLockHolder.

When HTMLMediaElement is notified that its ScriptExecutionContext has been destroyed, also
clear the list of pending Promises.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::contextDestroyed):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202923 268f45cc-cd09-0410-ab3c-d52691b4dbfc