Expose crossOrigin attribute as a static property in HTMLMediaElement https://bugs.webkit.org/show_bug.cgi?id=159459
Patch by Olivier Blin <olivier.blin@softathome.com> on 2016-07-08
Reviewed by Chris Dumez.

The crossOrigin attribute is already used for MediaResourceLoader
(r119742 and r175050), but it was not exposed as a static property.

This fixes VR360 support in Dailymotion, since it uses the "in"
operator to detect if crossOrigin is supported by the
HTMLVideoElement, in order to enable VR360.

No new tests, rebaselined existing tests, 150 WPT tests are fixed.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::setCrossOrigin):
(WebCore::HTMLMediaElement::crossOrigin):
* html/HTMLMediaElement.h:
* html/HTMLMediaElement.idl:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202980 268f45cc-cd09-0410-ab3c-d52691b4dbfc