Object.defineProperty() should maintain existing getter / setter if not overridden in the new descriptor https://bugs.webkit.org/show_bug.cgi?id=159576 <rdar://problem/27242197>
Reviewed by Mark Lam.

Source/JavaScriptCore:

Object.defineProperty() should maintain existing getter / setter if not
overridden in the new descriptor. Previously, if the property is a had
a custom getter / setter, and if the new descriptor only had a setter
(or only a getter), JSC would clear the existing getter (or setter).
This behavior did not match the EcmaScript specification or Firefox /
Chrome. This patch fixes the issue.

This fixes searching and search suggestions on www.iciba.com.

* runtime/JSObject.cpp:
(JSC::validateAndApplyPropertyDescriptor):

LayoutTests:

Add test coverage.

* js/dom/customAccessor-defineProperty-expected.txt: Added.
* js/dom/customAccessor-defineProperty.html: Added.
* js/dom/customAccessor-defineProperty_getterOnly-expected.txt: Added.
* js/dom/customAccessor-defineProperty_getterOnly.html: Added.
* js/dom/customAccessor-defineProperty_setterOnly-expected.txt: Added.
* js/dom/customAccessor-defineProperty_setterOnly.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203004 268f45cc-cd09-0410-ab3c-d52691b4dbfc