Cleanup: Remove use of PassRefPtr from class HTMLTableElement https://bugs.webkit.org/show_bug.cgi?id=159587
Reviewed by Chris Dumez.

* html/HTMLTableElement.cpp:
(WebCore::HTMLTableElement::setCaption): Take a rvalue reference to a RefPtr instead of a PassRefPtr.
(WebCore::HTMLTableElement::setTHead): Take a rvalue reference to a RefPtr instead of a PassRefPtr. Also
fix a style nit; add curly braces around the for-loop body since its body is more than a single line.
(WebCore::HTMLTableElement::createTHead): Use Ref::copyRef() instead of Ref::ptr() to pass the instantiated
table section to better convey that we are passing a copy of the table section.
(WebCore::HTMLTableElement::createCaption): Ditto.
* html/HTMLTableElement.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203016 268f45cc-cd09-0410-ab3c-d52691b4dbfc