Red and blue colors are swapped in video rendered through WebGL when GSTREAMER_GL is enabled https://bugs.webkit.org/show_bug.cgi?id=159621
Patch by Olivier Blin <olivier.blin@softathome.com> on 2016-07-11
Reviewed by Philippe Normand.

When a video is rendered through WebGL, and GSTREAMER_GL is enabled, red and blue colors are swapped.
This occurs for example with the following videos:
http://www.scirra.com/labs/bugs/webglvideo/
http://www.dailymotion.com/embed/video/x4jiicp?autoplay=1

This is because ImageGStreamerCairo expects video frames in either
BGRA or ARGB, while when GSTREAMER_GL is enabled,
createVideoSinkGL() forces a RGBA format.

Without GSTREAMER_GL, the rendering is fine since
VideoSinkGStreamer uses either BGRA or ARGB.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamerBase.cpp:
(WebCore::MediaPlayerPrivateGStreamerBase::createVideoSinkGL):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203059 268f45cc-cd09-0410-ab3c-d52691b4dbfc