Honor the second argument to FontFaceSet.load and FontFaceSet.check https://bugs.webkit.org/show_bug.cgi?id=159607 <rdar://problem/27284902>
Reviewed by Zalan Bujtas.

Source/WebCore:

This second argument is used in conjunction with the unicode-range CSS property, so that
loading from a FontFaceSet only loads the fonts which actually match the characters given.
Previously, we hadn't implemented proper support for this unicode-range property, but now
that we have implemented it, we should honor this second argument.

Test: fast/text/unicode-range-javascript.html

* css/CSSFontFace.cpp:
(WebCore::CSSFontFace::rangesMatchCodePoint):
* css/CSSFontFace.h:
* css/CSSFontFaceSet.cpp:
(WebCore::codePointsFromString):
(WebCore::CSSFontFaceSet::matchingFaces):

LayoutTests:

* fast/text/font-face-set-document-expected.txt:
* fast/text/font-face-set-document.html:
* fast/text/font-face-set-javascript-expected.txt:
* fast/text/font-face-set-javascript.html:
* fast/text/unicode-range-javascript-expected.txt: Added.
* fast/text/unicode-range-javascript.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203092 268f45cc-cd09-0410-ab3c-d52691b4dbfc