[WK2][iOS] Intermittent crash in [UIApplication beginBackgroundTaskWithName] expiration handler https://bugs.webkit.org/show_bug.cgi?id=159648 <rdar://problem/27219361>
Reviewed by Anders Carlsson.

We see an intermittent crash in [UIApplication beginBackgroundTaskWithName] expiration handler.
From the traces, we noticed the expiration handler can get called on a non-main thread under
certain conditions, which our code is designed to deal with.

In particular, we get called on a non-main thread when UIKit fails to acquire the assertion.
Update our expiration handler to always notify the clients on the main thread.

* UIProcess/ios/ProcessAssertionIOS.mm:
(-[WKProcessAssertionBackgroundTaskManager _updateBackgroundTask]):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203094 268f45cc-cd09-0410-ab3c-d52691b4dbfc