REGRESSION (202694): Audio and Video playback controls: Cannot find a position slider to adjust playback position using VO. https://bugs.webkit.org/show_bug.cgi?id=159661 <rdar://problem/27285135>
Reviewed by Eric Carlson.

Source/WebCore:

The change in r202694 caused MediaDocuments to not always
show their scrubber. The fix is to reduce the minimum amount
of size needed to show the scrubber.

Test: media/controls/default-size-should-show-scrubber.html

* Modules/mediacontrols/mediaControlsApple.js: 80 pixels is enough
to show the scrubber.

LayoutTests:

Add a test that checks the scrubber is visible on a default media element.
Unfortunately this isn't quite an accurate test, since we really want to
examine a Media Document, but it's close enough.

* media/controls/default-size-should-show-scrubber-expected.txt: Added.
* media/controls/default-size-should-show-scrubber.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203120 268f45cc-cd09-0410-ab3c-d52691b4dbfc