[GLib] Use a GSource instead of a thread to poll memory pressure eventFD in linux implementation https://bugs.webkit.org/show_bug.cgi?id=159346
Reviewed by Antonio Gomes.

The eventFD file descriptor is pollable, so it would be much better to use a poll instead of a blocking read in
a secondary thread and then communicate back to the main thread. This is very easy to do with GSource in GLib,
so we could use that when GLib is available and keep the current implementation as a fallback.

* platform/MemoryPressureHandler.cpp:
(WebCore::m_holdOffTimer): Use a RunLoop timer.
* platform/MemoryPressureHandler.h:
* platform/linux/MemoryPressureHandlerLinux.cpp:
(WebCore::MemoryPressureHandler::EventFDPoller::EventFDPoller): Helper class do the eventFD polling.
(WebCore::MemoryPressureHandler::logErrorAndCloseFDs): Check if file descriptors are -1 not 0.
(WebCore::MemoryPressureHandler::install): Return early also if the hold off timer is active. Use EventFDPoller
to do the polling.
(WebCore::MemoryPressureHandler::uninstall): Stop the hold off timer and clear the EventFDPoller.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@203216 268f45cc-cd09-0410-ab3c-d52691b4dbfc